+++ nflow-tests/src/test/java/io/nflow/tests/CronWorkflowTest.java
@@ -54,8 +54,8 @@ public class CronWorkflowTest extends AbstractNflowTest {
 
   @Test
   @Order(2)
-  public void letItRunFor5Seconds() throws InterruptedException {
-    SECONDS.sleep(5);
+  public void letItRunForTenSeconds() throws InterruptedException {
+    SECONDS.sleep(10);
   }
 
   @Test
@@ -63,8 +63,10 @@ public class CronWorkflowTest extends AbstractNflowTest {
   public void verifyItHasRunPeriodically() {
     List<Action> actions = getWorkflowInstance(resp.id).actions;
     long scheduleActions = actions.stream().filter(a -> CronWorkflow.State.schedule.name().equals(a.state)).count();
+    long waitActions = actions.stream().filter(a -> CronWorkflow.State.waitForWorkToFinish.name().equals(a.state)).count();
     long doWorkActions = actions.stream().filter(a -> CronWorkflow.State.doWork.name().equals(a.state)).count();
     assertThat(scheduleActions, is(greaterThanOrEqualTo(1L)));
+    assertThat(waitActions, is(greaterThanOrEqualTo(1L)));
     assertThat(doWorkActions, is(greaterThanOrEqualTo(1L)));
   }